Shift all files into Duke package
