+++ src/main/java/duke/Command.java
@@ -1,3 +1,5 @@
+package duke;
+
 public abstract class Command {
     public abstract void execute(TaskList tasks, Storage storage, Ui ui);
     public boolean isExit() {
