+++ src/main/java/duke/DeleteCommand.java
@@ -1,3 +1,5 @@
+package duke;
+
 public class DeleteCommand extends Command {
     private int taskIndex;
