+++ src/main/java/duke/ExitCommand.java
@@ -1,3 +1,5 @@
+package duke;
+
 public class ExitCommand extends Command {
     @Override
     public void execute(TaskList tasks, Storage storage, Ui ui) {
