+++ src/main/java/duke/OccurringCommand.java
@@ -1,3 +1,5 @@
+package duke;
+
 import java.time.LocalDate;
 
 public class OccurringCommand extends Command {
